cron/requirements.txt
newspaper3k==0.2.8
pandas==2.2.2
pendulum==3.0.0
requests==2.32.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sentence-transformers==3.0.1
umap-learn==0.5.6
scikit-learn==1.5.1
numpy==1.26.4
supabase==2.5.0